Refactoring: move the extension to a separate file
