+++ src/xcode/ENA/ENA/Source/Services/Risk/Calculation/__tests__/RiskCalculationTests.swift
@@ -21,7 +21,7 @@
 import ExposureNotification
 import XCTest
 
-final class RiskCalculationTests: XCTestCase {
+class RiskCalculationTests: XCTestCase {
 
 	private let store = MockTestStore()
 
@@ -546,98 +546,5 @@ final class RiskCalculationTests: XCTestCase {
 		)
 		// The risk level did not change - we only care about changes between low and increased
 		XCTAssertFalse(risk?.riskLevelHasChanged ?? true)
-	}
-}
-
-// MARK: - Helpers
-
-private extension RiskCalculationTests {
-
-	private var appConfig: SAP_ApplicationConfiguration {
-		makeAppConfig(w_low: 1.0, w_med: 0.5, w_high: 0.5)
-	}
-
-	private var summaryLow: CodableExposureDetectionSummary {
-		makeExposureSummaryContainer(maxRiskScoreFullRange: 80, ad_low: 10, ad_mid: 10, ad_high: 10)
-	}
-
-	private var summaryMed: CodableExposureDetectionSummary {
-		makeExposureSummaryContainer(maxRiskScoreFullRange: 128, ad_low: 15, ad_mid: 15, ad_high: 15)
-	}
-
-	private var summaryHigh: CodableExposureDetectionSummary {
-		makeExposureSummaryContainer(maxRiskScoreFullRange: 255, ad_low: 30, ad_mid: 30, ad_high: 30)
-	}
-
-	enum PreconditionState {
-		case valid
-		case invalid
-	}
-
-	private func preconditions(_ state: PreconditionState) -> ExposureManagerState {
-		switch state {
-		case .valid:
-			return .init(
-				authorized: true,
-				enabled: true,
-				status:
-				.active
-			)
-		default:
-			return .init(authorized: true, enabled: false, status: .disabled)
-		}
-	}
-
-	private func makeExposureSummaryContainer(
-		maxRiskScoreFullRange: Int,
-		ad_low: Double,
-		ad_mid: Double,
-		ad_high: Double
-	) -> CodableExposureDetectionSummary {
-		.init(
-			daysSinceLastExposure: 0,
-			matchedKeyCount: 0,
-			maximumRiskScore: 0,
-			attenuationDurations: [ad_low, ad_mid, ad_high],
-			maximumRiskScoreFullRange: maxRiskScoreFullRange
-		)
-	}
-
-	/// Makes an mock `SAP_ApplicationConfiguration`
-	///
-	/// Some defaults are applied for ad_norm, w4, and low & high ranges
-	private func makeAppConfig(
-		ad_norm: Int32 = 25,
-		w4: Int32 = 0,
-		w_low: Double,
-		w_med: Double,
-		w_high: Double,
-		riskRangeLow: ClosedRange<Int32> = 1...5,
-		// Gap between the ranges is on purpose, this is an edge case to test
-		riskRangeHigh: Range<Int32> = 6..<11
-	) -> SAP_ApplicationConfiguration {
-		var config = SAP_ApplicationConfiguration()
-		config.attenuationDuration.defaultBucketOffset = w4
-		config.attenuationDuration.riskScoreNormalizationDivisor = ad_norm
-		config.attenuationDuration.weights.low = w_low
-		config.attenuationDuration.weights.mid = w_med
-		config.attenuationDuration.weights.high = w_high
-
-		var riskScoreClassLow = SAP_RiskScoreClass()
-		riskScoreClassLow.label = "LOW"
-		riskScoreClassLow.min = riskRangeLow.lowerBound
-		riskScoreClassLow.max = riskRangeLow.upperBound
-
-		var riskScoreClassHigh = SAP_RiskScoreClass()
-		riskScoreClassHigh.label = "HIGH"
-		riskScoreClassHigh.min = riskRangeHigh.lowerBound
-		riskScoreClassHigh.max = riskRangeHigh.upperBound
-
-		config.riskScoreClasses.riskClasses = [
-			riskScoreClassLow,
-			riskScoreClassHigh
-		]
-
-		return config
 	}
 }
